requirements-test.txt
pytest==6.0.1
pytest-aiohttp==0.3.0
